# Initialize ABI functions lazily