regexp provides a large API to satisfy many use cases, but code will commonly only exercise a small portion of the surface, meaning only a small portion of our wazero ABI. Because each initialized function costs a relatively expensive call stack, we can reduce memory usage for such cases by initializing as needed. Even if all of the ABI is exercised, the negative performance impact should not be significant given we already have a mutex guarding, so there is no extra safety needed and it's a simple check.

After
```
goos: darwin
goarch: arm64
pkg: github.com/wasilibs/go-re2
BenchmarkCompile
BenchmarkCompile/Onepass
BenchmarkCompile/Onepass-10         	   18510	     61832 ns/op	  296924 B/op	      88 allocs/op
BenchmarkCompile/Medium
BenchmarkCompile/Medium-10          	   14674	     79494 ns/op	  296926 B/op	      88 allocs/op
BenchmarkCompile/Hard
BenchmarkCompile/Hard-10            	    3724	    317359 ns/op	  296882 B/op	      88 allocs/op
PASS
```

Before
```
goos: darwin
goarch: arm64
pkg: github.com/wasilibs/go-re2
BenchmarkCompile
BenchmarkCompile/Onepass
BenchmarkCompile/Onepass-10         	   16719	     68467 ns/op	  352925 B/op	     122 allocs/op
BenchmarkCompile/Medium
BenchmarkCompile/Medium-10          	   13615	     86617 ns/op	  352925 B/op	     121 allocs/op
BenchmarkCompile/Hard
BenchmarkCompile/Hard-10            	    3567	    324546 ns/op	  352930 B/op	     122 allocs/op
PASS
```