@mathetake 2023-03-29
https://github.com/wasilibs/go-re2/pull/43